Growformer Production QA Updates

+++ src/neuron.rs
@@ -147,7 +147,7 @@ impl Neuron {
 }
 
 pub fn sigmoid(x: f32) -> f32 {
-    1.0 / (1.0 + (-x).exp())
+    (1.0 / (1.0 + (-(x as f64)).exp())) as f32
 }
 
 pub fn sigmoid_derivative(x: f32) -> f32 {
